Add files via upload
Helen's original codes from her work on "A Successful Bite? Insights into Skin Distribution and Onward Transmission of Leishmania Parasites from Infected Mice to Sandflies".
The files 'EpidemicSS', 'GillepsieSS' and 'SIRmodelSS' implement a superspreading event which was from an assignment in the Modelling with Matlab module. This superspreading idea might be helpful if we take the idea into the project, thereby I shall have something to work with. Note that this model uses the Gillepsie algorithm.